route rcp logs to local5
Makes changes to a big variety of files in order to comprise logging
capability, and also take advantage of `syslog` for doing so.
logs were going not to the parent process, but to the system logger,
which is not optimal for analysis.
also guarantees log init is being called, which doesn't happen at
all.